+++ src/main/java/com/indref/industrial_reforged/api/tiers/FireboxTier.java
@@ -1,5 +1,5 @@
 package com.indref.industrial_reforged.api.tiers;
 
 public interface FireboxTier {
-    int getMaxHeatOutput();
+    float getMaxHeatOutput();
 }
